pull out generic table functionality
