+++ src/frontend/src/components/Table/style.ts
@@ -1,5 +1,24 @@
 import styled from "@emotion/styled";
-import { getColors, TableRow } from "czifui";
+import { getColors, getSpaces, InputCheckbox, TableRow } from "czifui";
+
+// needed to keep search bar sticky
+export const StyledWrapper = styled.div`
+  flex: 1 1 auto;
+  overflow-y: auto;
+
+  & > div {
+    overflow: auto;
+  }
+`;
+
+export const StyledInputCheckbox = styled(InputCheckbox)`
+  ${(props) => {
+    const spaces = getSpaces(props);
+    return `
+      padding-bottom: ${spaces?.xxs}px;
+    `;
+  }}
+`;
 
 export const StyledTableRow = styled(TableRow)`
   ${(props) => {
